requirements.txt
asgiref==3.8.1
chardet==5.2.0
Django==5.1.5
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
exceptiongroup==1.2.2
inflection==0.5.1
iniconfig==2.1.0
mysqlclient==2.2.7
numpy==2.2.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pluggy==1.5.0
PyJWT==2.10.1
pytest==8.3.5
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
reportlab==4.4.0
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
